refactor(flist): encapsulate generator sort behind DualFileList API
Replace direct `as_mut_vec()` access in the generator's two sort sites
with `DualFileList::sort_with_parallel`, which owns the indirect-sort +
cycle-following permutation. This removes two of the legacy-Vec
back-doors that block dropping `Vec<FileEntry>` from the dual
representation.

- Move `apply_permutation_in_place` from `transfer::generator::file_list`
  to `protocol::flist::sort` as a public helper; re-export through
  `protocol::flist`. The existing transfer-side tests reach it via the
  same `apply_permutation_in_place` symbol through a `pub(super) use`
  re-export so the test surface is unchanged.
- Add `DualFileList::sort_with_parallel<P>` that builds an index
  permutation via `compare_file_entries`, then applies it to both the
  legacy `Vec<FileEntry>` and a caller-owned parallel slice (the
  generator's `full_paths`) in lockstep. Sorts only the legacy Vec for
  now; the shadow `FlatFileList` is never read in production so
  reordering it is deferred to a follow-up that drops the legacy field.
- Migrate `build_file_list` and `build_file_list_with_base` sort blocks
  to the new method. Removes the local `compare_file_entries` import
  and the embedded permutation helper from `generator/file_list/mod.rs`.

No behavior change: the new method runs the same comparator, same
qsort/stable choice, and the same cycle-following permutation as the
original inline code.

Step 1 of 4 toward dropping the legacy `Vec<FileEntry>` from
`DualFileList` (RSS-A.LAND / FP344-5).
